// src/data/class.xsd.d.ts
//generated file

import type {Attr} from '../pages/classes/components/Attr'

export type constant = {
	name: Attr
	value: Attr
	enum?: Attr
	is_bitfield?: Attr
	is_deprecated?: Attr
	is_experimental?: Attr
}
export type member = {
	name: Attr
	type: Attr
	setter: Attr
	getter: Attr
	overrides?: Attr
	enum?: Attr
	default?: Attr
	is_deprecated?: Attr
	is_experimental?: Attr
}
export type theme_item = {
	name: Attr
	data_type: Attr
	type: Attr
	default?: Attr
}
export type tutorial = {
	title?: Attr
}
export type constructor_ = {
	name?: Attr
	qualifiers?: Attr
}
export type constructor_return = {
	type: Attr
	enum?: Attr
}
export type constructor_param = {
	index: Attr
	name: Attr
	type: Attr
	enum?: Attr
	default?: Attr
}
export type method = {
	name?: Attr
	qualifiers?: Attr
	is_deprecated?: Attr
	is_experimental?: Attr
}
export type method_return = {
	type: Attr
	enum?: Attr
}
export type method_returns_error = {
	number: Attr
}
export type method_param = {
	index: Attr
	name: Attr
	type: Attr
	enum?: Attr
	default?: Attr
}
export type signal = {
	name?: Attr
}
export type signal_param = {
	index: Attr
	name: Attr
	type: Attr
}
export type annotation = {
	name?: Attr
	qualifiers?: Attr
}
export type annotation_return = {
	type: Attr
	enum?: Attr
}
export type annotation_param = {
	index: Attr
	name: Attr
	type: Attr
	enum?: Attr
	default?: Attr
}
export type operator = {
	name?: Attr
	qualifiers?: Attr
}
export type operator_return = {
	type: Attr
	enum?: Attr
}
export type operator_param = {
	index: Attr
	name: Attr
	type: Attr
	enum?: Attr
	default?: Attr
}